+++ CHANGELOG.md
@@ -25,7 +25,7 @@ The format is based on [Keep a Changelog].
   order to retrieve the specified number of jobs, if higher. (\#263)
 - The `backend.properties()` function now accepts an optional `datetime_filter` 
   parameter. If specified, the function returns the backend properties closest 
-  to, but less than, the specified datetime filter.
+  to, but less than, the specified datetime filter (\#277).
 
 ### Fixed